fix: Fix handling of the SIGUSR1 on Windows
